test: move tests needing fixtures into tests/
