+++ tests/test_simplification.rs
